Adds password authentication to the api
